Added Admin Page: 'app.py' has new endpoints, admin_dashboard, update_user, and admin_delete_user. 'admin_dashboard.html' provides a paginated user table with username links to a modal where first_name, last_name, e-mail adddress can be updated, the user toggled inactive/active and the user can be permanently deleted.
Styled 'edit_user_profile.html' and 'members.html' for text+icon placement consistency in the delete modals.

Increased the collapsible-body background whiteness to 0.35 from 0.3 to increase the black-text-to-white-background contrast ratio.

Turned on the Flask Debug Toolbar while working on the Admin Page to validate the form fields and URL variables passed are correct.

Added measures to both 'admin_dashboard.html' and 'app.py' to enusre that the admin user is not set to inactive, nor deleted even if the  admin id is known.