In 'config.py' "USER_ENABLE_CHANGE_USERNAME = False" as a username change (Flask-Login) will not update the relationship to the user's books. * Future feature *
Allowing for users with the role of Admin to access the Admin Dashboard, changed {% if current_user.username == "admin"%} to {% if "Admin" in current_user.roles %} in 'admin_dashboard.html'. Tested successfully with user gaff.

Shortened the labels in the user profile modal. Complies with UX Design of Book Repository (all other form field labels).

Added a try-except to the loading of 'genre.json', in the event it doesn't exist, load_genres and home_page in 'app.py'.

Updated '.gitignore' to ignore rotating log files.