Updated the help tooltip for the Admin Dashboard to include information about the Statistics modal in 'admin_dashboard.html'.
Updated 'book.json' by adding further books, and 'genre.json' by removing the icon field as it's not used. Formatted both to be more human readable. 'genre.json' is loaded the first time the home/landing page is accessed, together with the creation of the admin user. Both 'genre.json' and 'book.json' can be loaded in the Admin Dashboard too.

Configured the loggging in 'app.py': rotating file logging in my local development environment, and on Heroku it's the application logger. Will review the app.logger statements in 'app.py' to make sure they all follow the same format.

Added additional routes to the home/landing page, "/", "/index", and "index.html" to cover the most common ways to access a website. Moved the book objects to a file, 'book.json'.

Having both 'genre.json' and 'book.json' as files make it possible for an admin to update both files in future.