Updated delete_book in 'app.py' to ensure an unauthenticated or authenticated though not owner of the book can delete the book using a URL (http://127.0.0.1:5000/delete_book/5f453c83698a401d389d56a8).
Testing: signed in (authenticated) though not owner of the book will warn the user not the 'act the maggot' by trying to delete a book not belonging to them. Unauthenticated external user will be asked to sign in to delete the book which will not be possible. Unauthenticated user and owner of the book will require to sign in, and on successful authentication will delete the book.